Someone made code that didn't compile
I can't fix it because I don't know what the code's intended purpose is.  I have to leave it to tomorrow (the last day we can do work) for someone to figure out why the code is causing an error.  I know what the error is, but I don't know how to fix it so that it will act in the way that the original creator intended it for.  Besides, I have too much work due tomorrow to care.